requirements.txt
Flask>=2.3.0,<3.1.0
requests>=2.28.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<2.1.0
PyYAML>=6.0.0,<7.0.0
pybit>=5.8.0
flask-socketio>=5.3.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
websocket-client>=1.6.0,<2.0.0
psutil>=5.9.0
ta>=0.10.0
eventlet>=0.33.0
gunicorn>=21.0.0 
